+++ src/Components/Main/Pages/NewsPage/NewsPage.module.scss
